+++ libs/components/src/lib/components/table/td/td.component.ts
@@ -1,6 +1,15 @@
-import { ChangeDetectionStrategy, Component, ContentChild, HostBinding, Input } from '@angular/core';
+import {
+  ChangeDetectionStrategy,
+  Component,
+  ContentChild,
+  HostBinding,
+  Input,
+  OnDestroy,
+  OnInit,
+} from '@angular/core';
 import { NgControl } from '@angular/forms';
 import { PrizmTableCellStatus } from '../table.types';
+import { PrizmTdService } from './td.service';
 
 @Component({
   // eslint-disable-next-line @angular-eslint/component-selector
@@ -9,10 +18,24 @@ import { PrizmTableCellStatus } from '../table.types';
   styleUrls: [`./td.component.less`],
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class PrizmTdComponent {
+export class PrizmTdComponent implements OnInit, OnDestroy {
   @Input() @HostBinding('attr.status') public status: PrizmTableCellStatus = 'default';
+
+  @Input()
+  @HostBinding('attr.colspan')
+  colspan: string | number = 1;
 
   @HostBinding(`class._editable`)
   @ContentChild(NgControl)
   readonly control: unknown;
+
+  constructor(private readonly tdService: PrizmTdService) {}
+
+  public ngOnInit(): void {
+    this.tdService.increment(parseInt(this.colspan.toString(), 10));
+  }
+
+  public ngOnDestroy(): void {
+    this.tdService.decrement(parseInt(this.colspan.toString(), 10));
+  }
 }