+++ apps/doc/src/app/about-prizm/changelog/CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file.
 
 ### Bugfixes
 
+- fix(components/navigation): select current item #477
 - fix(doc): anchor links do not work on pages that have not loaded #560
 
 ## [1.4.1](https://github.com/zyfra/Prizm) (31-07-2023)