+++ apps/doc/src/app/about-prizm/changelog/CHANGELOG.md
@@ -2,7 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
-## [2.1.4-next.1](https://github.com/zyfra/Prizm) (01-08-2023)
+## [2.1.4-next.2](https://github.com/zyfra/Prizm) (03-08-2023)
+
+### Features
+
+- feat(components/table): add support dynamic amount of columns in group #259
+
+## [2.1.4-next.1](https://github.com/zyfra/Prizm) (02-08-2023)
 
 ### Features